Move service provider binding functionality to trait
